Tasks get associated with workflowStateId of a different workspace (#163)
in the completeTask API. Because of this, when tasks are marked
done from the client side form  a different workspace such as
(ProdAri) then the task would disapear since it would get
workflowStateId of a different workspace. There was a small
one liner miss in the db query which is added in this commit.